+++ src/textual/command.py
@@ -458,6 +458,8 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
         """The command that was selected by the user."""
         self._busy_timer: Timer | None = None
         """Keeps track of if there's a busy indication timer in effect."""
+        self._no_matches_timer: Timer | None = None
+        """Keeps track of if there are 'No matches found' message waiting to be displayed."""
         self._providers: list[Provider] = []
         """List of Provider instances involved in searches."""
 
@@ -558,6 +560,38 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
                 self._show_busy = True
 
         self._busy_timer = self.set_timer(self._BUSY_COUNTDOWN, _become_busy)
+
+    def _stop_no_matches_countdown(self) -> None:
+        """Stop any 'No matches' countdown that's in effect."""
+        if self._no_matches_timer is not None:
+            self._no_matches_timer.stop()
+            self._no_matches_timer = None
+
+    _NO_MATCHES_COUNTDOWN: Final[float] = 0.5
+    """How many seconds to wait before showing 'No matches found'."""
+
+    def _start_no_matches_countdown(self) -> None:
+        """Start a countdown to showing that there are no matches for the query.
+
+        Adds a 'No matches found' option to the command list after `_NO_MATCHES_COUNTDOWN` seconds.
+        """
+        self._stop_no_matches_countdown()
+
+        def _show_no_matches() -> None:
+            command_list = self.query_one(CommandList)
+            command_list.add_option(
+                Option(
+                    Align.center(Text("No matches found")),
+                    disabled=True,
+                    id=self._NO_MATCHES,
+                )
+            )
+            self._list_visible = True
+
+        self._no_matches_timer = self.set_timer(
+            self._NO_MATCHES_COUNTDOWN,
+            _show_no_matches,
+        )
 
     def _watch__list_visible(self) -> None:
         """React to the list visible flag being toggled."""
@@ -732,6 +766,7 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
         command_list.clear_options().add_options(sorted(commands, reverse=True))
         if highlighted is not None:
             command_list.highlighted = command_list.get_option_index(highlighted.id)
+        self._list_visible = bool(command_list.option_count)
 
     _RESULT_BATCH_TIME: Final[float] = 0.25
     """How long to wait before adding commands to the command list."""
@@ -780,10 +815,7 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
         # grab a reference to that.
         worker = get_current_worker()
 
-        # We're ready to show results, ensure the list is visible.
-        self._list_visible = True
-
-        # Go into a busy mode.
+        # Reset busy mode.
         self._show_busy = False
 
         # A flag to keep track of if the current content of the command hit
@@ -861,13 +893,7 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
         # mean nothing was found. Give the user positive feedback to that
         # effect.
         if command_list.option_count == 0 and not worker.is_cancelled:
-            command_list.add_option(
-                Option(
-                    Align.center(Text("No matches found")),
-                    disabled=True,
-                    id=self._NO_MATCHES,
-                )
-            )
+            self._start_no_matches_countdown()
 
     @on(Input.Changed)
     def _input(self, event: Input.Changed) -> None:
@@ -878,6 +904,8 @@ class CommandPalette(ModalScreen[CallbackType], inherit_css=False):
         """
         event.stop()
         self.workers.cancel_all()
+        self._stop_no_matches_countdown()
+
         search_value = event.value.strip()
         if search_value:
             self._gather_commands(search_value)